==> analysis/02b_plots.Rmd <==
---
title: "02b_plots"
author: "Julia Pop"
date: "2025-09-22"
output: html_document
---

```{r setup, include=FALSE}
# code to troublesheet file path issues. 
# By default, knitr uses the folder containing the .Rmd file as
# the working directory. This code allows us to set the folder
# with our .rProj in it as the root. 

# find_root looks upwards from the .Rmd until it finds a file or
# folder that matches the criteria "folder with a .Rproj file in
# it."

knitr::opts_knit$set(root.dir = rprojroot::find_root(rprojroot::is_rstudio_project))
```

## Plot themes and packages
```{r, Echo = FALSE}
source("scripts/libraries.R")

#Customize theme 
#http://joeystanley.com/blog/custom-themes-in-ggplot2 
# theme_SJ <- function () { 
#  theme_base(base_size=8, base_family="PT Sans") %+replace% # base size scales everything
 #   theme(
 #     panel.background  = element_blank(), # back panel blank
 #     plot.background = element_rect(fill="white", colour=NA), # plot bgd white, with no border
 #     plot.title=element_text(face="plain",hjust=0), # 
 #     plot.subtitle = element_text(color="dimgrey", hjust=0, size=8), # hjust = left aligned
 #     axis.title=element_text(face="plain"), # face = plain means not bold
 #     panel.grid.minor=element_blank(), # removes minor grid lines
 #     strip.text.x = element_text(face="italic"),
 #     panel.grid.major=element_line(color="lightgrey"))}

# theme_set(theme_paleo())  # pre-set the bw theme
```

## Pull in data
```{r}
all_lakes <- read.csv("output/all_lakes.csv")

UFM_short <- read.csv("output/UFM_short.csv")
LFM_short <- read.csv("output/LFM_short.csv")
TRK_short <- read.csv("output/TRK_short.csv")

UFM_short_extrap <- read.csv("output/UFM_short_extrap.csv") %>%
  select(-X)
LFM_short_extrap <- read.csv("output/LFM_short_extrap.csv") %>%
  select(-X)
TRK_short <- read.csv("output/TRK_short.csv") %>%
  select(-X)
```

# ADM range table 
```{r}
# LFM max est age
LFM_short_extrap %>%
  slice_min(year_CE_plot) %>%
  pull(year_CE_plot)

# LFM max est depth
LFM_short_extrap %>%
  slice_min(year_CE_plot) %>%
  pull(depth_plot)

# LFM max reliably est age
LFM_short_extrap %>%
  filter(year_CE_plot > 1925) %>%
    slice_min(year_CE_plot) %>%
    pull(year_CE_plot)

# LFM max reliably est depth
LFM_short_extrap %>%
  filter(year_CE_plot > 1925) %>%
    slice_min(year_CE_plot) %>%
    pull(depth_plot)

# LFM min est age
LFM_short_extrap %>%
  slice_max(year_CE_plot) %>%
  pull(year_CE_plot)

# LFM min est depth
LFM_short_extrap %>%
  slice_max(year_CE_plot) %>%
  pull(depth_plot)

# UFM max est age
UFM_short_extrap %>% 
  slice_min(year_CE_plot) %>%
  pull(year_CE_plot)

# UFM max est depth
UFM_short_extrap %>% 
  slice_min(year_CE_plot) %>%
  pull(depth_plot)

# UFM min est age
UFM_short_extrap %>% 
  slice_max(year_CE_plot) %>%
  pull(year_CE_plot)

# UFM min est depth
UFM_short_extrap %>% 
  slice_max(year_CE_plot) %>%
  pull(depth_plot)

# TRK max est age
TRK_short %>%
  slice_min(year_CE_plot) %>%
  pull(year_CE_plot)

# TRK max est depth
TRK_short %>%
  slice_min(year_CE_plot) %>%
  pull(depth_plot)

# TRK min est age
TRK_short %>%
  slice_max(year_CE_plot) %>%
  pull(year_CE_plot)

# TRK min est depth
TRK_short %>%
  slice_max(year_CE_plot) %>%
  pull(depth_plot)

# UFM extrapolated age at bottom of disturbance (10.5cm)

UFM_short_extrap %>% 
  filter(depth_plot > 10.5) %>% 
  slice_max(year_plot_extrap) %>%
  pull(year_plot_extrap)

LFM_adm <- readRDS("output/LFM_adm.rds")
predict(LFM_adm, depth = 25)$age

TRK_short <- read.csv("output/TRK_short.csv") %>%
  select(-X) %>% 
  mutate(C.Nmolar = C.N*1.167, .after = C.N)

View(select(UFM_short_extrap, c(depth_plot, year_plot_extrap, year_CE_plot, percent.N)))
View(select(LFM_short_extrap, c(depth_plot, year_plot_extrap, year_CE_plot, percent.N)))
View(select(TRK_short, c(depth_plot, year_plot_extrap, year_CE_plot, d13C)))
names(TRK_short)

sum(!is.na(LFM_short_extrap$year_CE_plot))
sum(!is.na(UFM_short_extrap$year_CE_plot))
sum(!is.na(TRK_short$year_CE_plot))

TRK_short %>%
  slice_max(year_CE_plot)

TRK_adm <- readRDS("output/TRK_adm.rds")
predict(TRK_adm, depth = 10.5)$age

```


# ADM plots
## UFM
```{r}

# UFM -> 1860
# UFM domain: 1860 - 2025
  # solid-dashed cutoff: 1949.942

# crop to GAM domain:
data <- UFM_short_extrap %>% filter(year_plot_extrap > 1860)

# plot
UFM_adm_plot <- ggplot(data, aes(x = year_plot_extrap, y = depth_plot)) +
    geom_point(size = 1) +
   # dashed line where year < 1949
#  geom_line(
#    data = data %>% filter(year_plot_extrap < 1950),
#    linewidth = 0.4,
#    color = "black",
#    linetype = "dashed") +
  # solid line where year > 1949
  geom_line(
    data = data, 
    linewidth = 0.35,
    color = "black",
    linetype = "solid") +
  scale_y_reverse(limits = c(max(UFM_short_extrap$depth_plot), 0), expand = c(0, 0)) +
  scale_x_continuous(limits = c(1860, 2020),
                      breaks = seq(1860, 2020, by = 40),  # tick marks every 20 years
    # minor_breaks = seq(1860, 2020, by = 10),  # minor grid lines every 20 years
                      position = "bottom") +  # move x-axis to the bottom
  labs(
    x = "Year",               # name for x-axis
    y = "Depth (cm)",          # name for y-axis
  ) +
  ggtitle("UFM") + 
  theme_bw() +
  theme(
     axis.title.x = element_text(size = 9, colour = "black"),
     axis.title.y = element_text(size = 9, colour = "black"),
     axis.text.x  = element_text(size = 7.5, colour = "black"),
     axis.text.y  = element_text(size = 7.5, colour = "black"),
     plot.title = element_text(size = 10, face = "bold", colour = "black"),
  #  axis.title.x.top = element_text(vjust = 0),    # adjust title position
   # axis.text.x.top = element_text(vjust = 0),     # adjust tick labels
  #  axis.ticks.x.top = element_line(),             # show ticks
 #   axis.title.x.bottom = element_blank(),         # remove bottom axis
  #  axis.text.x.bottom = element_blank(),
   # axis.ticks.x.bottom = element_blank(),
    aspect.ratio = 1.5)  # change aspect ratio

UFM_adm_plot

# check against original 
UFM_adm <- readRDS("output/UFM_adm.rds")
plot(UFM_adm)

# save

ggsave("figures/UFM_adm_plot.png",
       dpi=600,
       height=3,
       width=2,
       units="in")

```

## TRK
```{r}

# TRK domain: 1910 (.75) - 2025
  # solid-dashed cutoff: 1982.209

# don't need to crop to GAM (dates don't go further back than 1860 anyway)
# data <- TRK_short %>% filter(year_plot_extrap > 1860)

# plot
TRK_adm_plot <- ggplot(TRK_short, aes(x = year_plot_extrap, y = depth_plot)) +
  # points: 
  geom_point(
    data = TRK_short,
    aes(x = year_plot_extrap, y = depth_plot),
    size = 1
  ) +
   # dashed line where year < 1982.209
 # geom_line(
#  data = TRK_short %>% filter(year_plot_extrap < 1988),
 #   linewidth = 0.4,
  #  color = "black",
   # linetype = "dotted") +
  # solid line where year > 1982.209
  geom_line(
    data = TRK_short, # %>% filter(year_plot_extrap > 1982),
    linewidth = 0.35,
    color = "black",
    linetype = "solid") +
  scale_y_reverse(limits = c(42, 0), expand = c(0, 0)) +
  scale_x_continuous(limits = c(1860, 2020),
                      breaks = seq(1860, 2020, by = 40),  # tick marks every 20 years
                   #  minor_breaks = seq(1860, 2020, by = 10),  
                   # minor grid lines every 20 years
                      position = "bottom") +  # move x-axis to the top
  labs(
    x = "Year",               # name for x-axis
    y = "Depth (cm)",          # name for y-axis
  ) +
  ggtitle("TRK") + 
  theme(
     axis.title.x = element_text(size = 9, colour = "black"),
     axis.title.y = element_text(size = 9, colour = "black"),
     axis.text.x  = element_text(size = 7.5, colour = "black"),
     axis.text.y  = element_text(size = 7.5, colour = "black"),
     plot.title = element_text(size = 10, face = "bold", colour = "black"),
  #  axis.title.x.top = element_text(vjust = 0),    # adjust title position
   # axis.text.x.top = element_text(vjust = 0),     # adjust tick labels
  #  axis.ticks.x.top = element_line(),             # show ticks
 #   axis.title.x.bottom = element_blank(),         # remove bottom axis
  #  axis.text.x.bottom = element_blank(),
   # axis.ticks.x.bottom = element_blank(),
    aspect.ratio = 1.5)  # change aspect ratio
   
TRK_adm_plot

# check against original 
TRK_adm <- readRDS("output/TRK_adm.rds")
plot(TRK_adm)

# save

ggsave("figures/TRK_adm_plot.png",
       dpi=600,
       height=3,
       width=2,
       units="in")

```

## LFM
```{r}

# LFM -> 1860
# LFM domain: 1860 - 2025
  # solid-dashed cutoff: none
  # period of confidence (rectangle): 1925
predict(LFM_adm, depth = 9)

# crop to GAM domain:
data <- LFM_short_extrap %>% filter(year_plot_extrap > 1860)

# plot
LFM_adm_plot <- ggplot(data, aes(x = year_plot_extrap, y = depth_plot)) +
  geom_point(size = 1) +
  geom_line(linewidth = 0.35,
    color = "black",
    linetype = "solid") +
  scale_y_reverse(limits = c(max(LFM_short_extrap$depth_plot), 0), expand = c(0, 0)) +
  scale_x_continuous(limits = c(1860, 2025),
                      breaks = seq(1860, 2020, by = 40),  # tick marks every 20 years
                     # minor_breaks = seq(1860, 2020, by = 10),  
                     # minor grid lines every 20 years
                      position = "bottom") +  # move x-axis to the top
  labs(
    x = "Year",               # name for x-axis
    y = "Depth (cm)",          # name for y-axis
  ) +
  ggtitle("LFM") + 
   theme(
     axis.title.x = element_text(size = 9, colour = "black"),
     axis.title.y = element_text(size = 9, colour = "black"),
     axis.text.x  = element_text(size = 7.5, colour = "black"),
     axis.text.y  = element_text(size = 7.5, colour = "black"),
     plot.title = element_text(size = 10, face = "bold", colour = "black"),
  #  axis.title.x.top = element_text(vjust = 0),    # adjust title position
   # axis.text.x.top = element_text(vjust = 0),     # adjust tick labels
  #  axis.ticks.x.top = element_line(),             # show ticks
 #   axis.title.x.bottom = element_blank(),         # remove bottom axis
  #  axis.text.x.bottom = element_blank(),
   # axis.ticks.x.bottom = element_blank(),
    aspect.ratio = 1.5)  # change aspect ratio

LFM_adm_plot

# check against original 
LFM_adm <- readRDS("output/LFM_adm.rds")
plot(LFM_adm)

# save

ggsave("figures/LFM_adm_plot.png",
       dpi=600,
       height=3,
       width=2,
       units="in")

```

# Plotting Preservation Indeces for Peter

## Pigment plots for Peter
```{r}

# pigment metric explanations
#         chla_pheoa = chl_a_nmol_gTC + pheo_a_nmol_gTC, 
#         chlb_pheob = chl_b_nmol_gTC + phaeo_b_nmol_gTC,
#         all_chl = chl_a_nmol_gTC + pheo_a_nmol_gTC + chl_ap_nmol_gTC, 
#         PI = chl_a_nmol_gTC/pheo_a_nmol_gTC,
#         AD_ratio = allo_nmol_gTC / diato_nmol_gTC, 
#         master_chl = (chl_a_nmol_gTC + chl_b_nmol_gTC) / (chl_ap_nmol_gTC + pheo_a_nmol_gTC + phaeo_b_nmol_gTC),
#         perc_chlb = (phaeo_b_nmol_gTC / pheo_a_nmol_gTC), 
#         perc_diatox = (diato_nmol_gTC/chla_pheoa)) %>%

# things to plot: 

# include: lake name, depth (plot?), year_CE
# plot: chl_a, pheo_a, all_chl, PI,  diatoxanthin, echineone (total cyanos)

UFM_short %>%
  pivot_longer(d13C:perc_diatox) %>%
  filter(name %in% c("chl_a_nmol_gTC", "pheo_a_nmol_gTC", 
                     "all_chl", "PI", 
                     "diato_nmol_gTC", "echine_nmol_gTC")) %>%
   mutate(name = recode(name, PI = "Preservation Index"), # change name of PI
         name = factor(name, levels = c(
           "chl_a_nmol_gTC",
           "pheo_a_nmol_gTC",
           "all_chl",
           "Preservation Index",
           "diato_nmol_gTC",
           "echine_nmol_gTC"
         ))) %>%
  filter(!is.na(value) & !is.na(depth_top)) %>% # remove missing points to make line continuous
  ggplot(aes(x=depth_top, y=value, color=name))+
  geom_point()+
  geom_line()+
  facet_wrap(.~name, scales="free_y")+
  theme(
    legend.position = "none",
    panel.background = element_rect(fill = "gray96", color = NA),  # gray panels
    panel.grid.major = element_line(color = "white", size = 0.5),  # visible major grid lines
    panel.grid.minor = element_blank(),  # remove minor grid lines
    plot.title = element_text(face = "bold", size = 10, hjust = 0, margin = margin(b=5)),  # bold, larger, centered, margin below
    strip.text.x = element_text(face = "italic", hjust = 0, size = 8)  # left-align facet titles
  ) + 
  scale_x_reverse() + 
  labs(
    title = "UFM Pigments",
    x = "Depth (cm)",
    y = expression(paste("Concentration (nmol ", g^{-1}, " TC)"))) + 
    scale_color_manual(values = c(
    "chl_a_nmol_gTC" = "#E69F00",      # orange
    "pheo_a_nmol_gTC" = "#56B4E9",     # sky blue
    "all_chl" = "#009E73",             # green
    "Preservation Index" = "gold2",  # yellow
    "diato_nmol_gTC" = "#0072B2",      # blue
    "echine_nmol_gTC" = "#D55E00"      # reddish-orange
  ))

LFM_short %>%
  pivot_longer(d13C:perc_diatox) %>%
  filter(name %in% c("chl_a_nmol_gTC", "pheo_a_nmol_gTC", 
                     "all_chl", "PI", 
                     "diato_nmol_gTC", "echine_nmol_gTC")) %>%
   mutate(name = recode(name, PI = "Preservation Index"), # change name of PI
         name = factor(name, levels = c(
           "chl_a_nmol_gTC",
           "pheo_a_nmol_gTC",
           "all_chl",
           "Preservation Index",
           "diato_nmol_gTC",
           "echine_nmol_gTC"
         ))) %>%
  filter(!is.na(value) & !is.na(depth_top)) %>% # remove missing points to make line continuous
  ggplot(aes(x=depth_top, y=value, color=name))+
  geom_point()+
  geom_line()+
  facet_wrap(.~name, scales="free_y")+
  theme(
    legend.position = "none",
    panel.background = element_rect(fill = "gray96", color = NA),  # gray panels
    panel.grid.major = element_line(color = "white", size = 0.5),  # visible major grid lines
    panel.grid.minor = element_blank(),  # remove minor grid lines
    plot.title = element_text(face = "bold", size = 10, hjust = 0, margin = margin(b=5)),  # bold, larger, centered, margin below
    strip.text.x = element_text(face = "italic", hjust = 0, size = 8)  # left-align facet titles
  ) + 
  scale_x_reverse() + 
  labs(
    title = "LFM Pigments",
    x = "Depth (cm)",
    y = expression(paste("Concentration (nmol ", g^{-1}, " TC)"))) + 
    scale_color_manual(values = c(
    "chl_a_nmol_gTC" = "#E69F00",      # orange
    "pheo_a_nmol_gTC" = "#56B4E9",     # sky blue
    "all_chl" = "#009E73",             # green
    "Preservation Index" = "gold2",  # yellow
    "diato_nmol_gTC" = "#0072B2",      # blue
    "echine_nmol_gTC" = "#D55E00"      # reddish-orange
  ))

TRK_short %>%
  pivot_longer(d13C:perc_diatox) %>%
  filter(name %in% c("chl_a_nmol_gTC", "pheo_a_nmol_gTC", 
                     "all_chl", "PI", 
                     "diato_nmol_gTC", "echine_nmol_gTC")) %>%
   mutate(name = recode(name, PI = "Preservation Index"), # change name of PI
         name = factor(name, levels = c(
           "chl_a_nmol_gTC",
           "pheo_a_nmol_gTC",
           "all_chl",
           "Preservation Index",
           "diato_nmol_gTC",
           "echine_nmol_gTC"
         ))) %>%
  filter(!is.na(value) & !is.na(depth_top)) %>% # remove missing points to make line continuous
  ggplot(aes(x=depth_top, y=value, color=name))+
  geom_point()+
  geom_line()+
  facet_wrap(.~name, scales="free_y")+
  theme(
    legend.position = "none",
    panel.background = element_rect(fill = "gray96", color = NA),  # gray panels
    panel.grid.major = element_line(color = "white", size = 0.5),  # visible major grid lines
    panel.grid.minor = element_blank(),  # remove minor grid lines
    plot.title = element_text(face = "bold", size = 10, hjust = 0, margin = margin(b=5)),  # bold, larger, centered, margin below
    strip.text.x = element_text(face = "italic", hjust = 0, size = 8)  # left-align facet titles
  ) + 
  scale_x_reverse() + 
  labs(
    title = "TRK Pigments",
    x = "Depth (cm)",
    y = expression(paste("Concentration (nmol ", g^{-1}, " TC)"))) + 
    scale_color_manual(values = c(
    "chl_a_nmol_gTC" = "#E69F00",      # orange
    "pheo_a_nmol_gTC" = "#56B4E9",     # sky blue
    "all_chl" = "#009E73",             # green
    "Preservation Index" = "gold2",  # yellow
    "diato_nmol_gTC" = "#0072B2",      # blue
    "echine_nmol_gTC" = "#D55E00"      # reddish-orange
  ))
```

## GREEBS Nutrient Plots (NOTE: No axis labels!)
```{r}
# Create nutrient plot
nutrient_plot <- all_lakes %>%
  pivot_longer(d13C:perc_diatox) %>%
  filter(name %in% c("d13C",
           "d15N",
           "percent.C",
           "percent.N",
           "C.N")) %>%
  mutate(name = recode(name, percent.C = "Total C (% dry mass)",   # change column names to change facet labels
                        percent.N = "N (% dry mass)",
                        C.N = "C:N ratio", 
                        d13C = "d13C (‰)",
                        d15N = "d15N (‰)"),
         name = factor(name, levels = c( # get plots in order
           "d13C (‰)",
           "d15N (‰)",
           "Total C (% dry mass)",
           "N (% dry mass)",
           "C:N ratio")),
         lake_name = factor(lake_name, levels = c("Upper Fourmile Lake", 
                                                  "Turkey Creek Lake", 
                                                  "Lower Fourmile Lake"))) %>% # draw lines in order
  filter(!is.na(value) & !is.na(depth_top)) %>% # remove missing points to make line continuous
  ggplot(aes(x=depth_top, y=value, color=lake_name))+
  labs(x = NULL, y = NULL)+
  geom_point()+
  geom_line(size = 1.2) + 
  facet_grid(name ~ ., scales = "free_y") + 
  theme(
    legend.position = "right",
    panel.background = element_rect(fill = "gray99", color = NA),  # gray panels
    panel.grid.major = element_blank(),
    panel.grid.minor = element_blank(),  # remove grid lines
    plot.title = element_text(face = "bold", size = 10, hjust = 0, margin = margin(b=5)),  # bold, larger, centered, margin below
    strip.text = element_blank(),  # remove default facet labels
    panel.border = element_rect(color = "black", fill = NA, linewidth = 1.2),  # black border
    axis.text.x = element_text(size = 16),   # x-axis ticks
    axis.text.y = element_text(size = 16),    # y-axis ticks
    legend.text  = element_text(size = 14),  # legend labels
  ) + 
  guides(color = guide_legend(override.aes = list(size = 2))) +
  scale_x_reverse() + 
  scale_color_manual( breaks = c("Upper Fourmile Lake", "Lower Fourmile Lake", "Turkey Creek Lake"), 
    values = c( 
    "Upper Fourmile Lake" = "#0072B2",
    "Lower Fourmile Lake" = "yellowgreen",
    "Turkey Creek Lake" = "#D55E00"
    ))

# save plots (with set dimensions)
ggsave("figures/nutrient_profiles.png", nutrient_plot, width = 8, height = 15, units = "in", dpi = 400) 

```

## GREEBS Pigment Plots
```{R}

# create a 4 x 2 plot!

pigment_order <- c(
  "chl_a_nmol_gTC", 
  "pheo_a_nmol_gTC", 
  "PI", 
  "lutein_nmol_gTC", 
  "diato_nmol_gTC", 
  "allo_nmol_gTC", 
  "echine_nmol_gTC", 
  "cantha_nmol_gTC"
)

# Create pigment plot
pigment_plot_GREEBS <- all_lakes %>%
  pivot_longer(d13C:perc_diatox) %>%
  filter(name %in% c("chl_a_nmol_gTC", "pheo_a_nmol_gTC", "PI", "lutein_nmol_gTC", 
                     "diato_nmol_gTC", "allo_nmol_gTC", "echine_nmol_gTC", "cantha_nmol_gTC"
  )) %>%
  filter(!is.na(value) & !is.na(depth_top)) %>% # remove missing points to make line continuous
  mutate(name = factor(name, levels = pigment_order), # set facet order
         lake_name = factor(lake_name, levels = c("Upper Fourmile Lake", 
                                                  "Turkey Creek Lake", 
                                                  "Lower Fourmile Lake"))) %>% # draw lines in order 
  ggplot(aes(x=depth_top, y=value, color=lake_name))+
  labs(x = NULL, y = NULL)+
  geom_point()+
  geom_line(size = 1.2) + 
  facet_wrap(~ name, scales = "free_y", nrow = 2, ncol = 4) +   # 2 rows, 4 columns
  theme(
    legend.position = "none", # remove legend (saved in figures folder)
    panel.background = element_rect(fill = "gray99", color = NA),  # gray panels
    panel.grid.major = element_blank(),
    panel.grid.minor = element_blank(),  # remove grid lines
    plot.title = element_text(face = "bold", size = 10, hjust = 0, margin = margin(b=5)),  # bold, larger, centered, margin below
    strip.text = element_blank(),  # remove default facet labels
    panel.border = element_rect(color = "black", fill = NA, size = 0.8),  # black border
    axis.text.x = element_text(size = 14),   # x-axis ticks
    axis.text.y = element_text(size = 14),    # y-axis ticks
    panel.spacing.x = unit(0.2, "in"),   # increase space between columns
    panel.spacing.y = unit(0.7, "in"),
  ) + 
  scale_x_reverse() +  # top ticks, no labels
  scale_color_manual( 
    values = c( 
    "Upper Fourmile Lake" = "#0072B2",  # 
    "Lower Fourmile Lake" = "yellowgreen",  # 
    "Turkey Creek Lake" = "#D55E00"))

# save plots (with set dimensions)
ggsave("figures/GREEBS_pigments.png", pigment_plot_GREEBS, width = 14, height = 5, units = "in", dpi = 300) 

# has no units, will go back and add those manually (can check quickly by deleting "strip.text = ..." line)

```

##Pigment flux plots vs. time
```{R}

# create a 4 x 2 plot!

pigment_flux_order <- c(
  "chl_a_flux", 
  "pheo_a_flux", 
  "PI", 
  "lutein_flux", 
  "diato_flux", 
  "allo_flux", 
  "echine_flux", 
  "cantha_flux"
)

# Create pigment plot
pigment_flux_plot <- 
  all_lakes %>%
  pivot_longer(cols = c(bd:year_CE_top, year_CE_bottom:s), names_to = "name", values_to = "value") %>%
  filter(name %in% c(
    "chl_a_flux", "pheo_a_flux", "PI", "lutein_flux",
    "diato_flux", "allo_flux", "echine_flux", "cantha_flux"
  )) %>% # filter instead of select bc these are rows after using pivot...
  # filter(!is.na(value) & !is.na(depth_top)) %>% # remove missing points to make line continuous
  mutate(name = factor(name, levels = pigment_flux_order),   # set facet order
         lake_name = factor(lake_name, levels = c("Upper Fourmile Lake", 
                                                  "Turkey Creek Lake", 
                                                  "Lower Fourmile Lake"))) %>% # draw lines in order
  ggplot(aes(x=year_CE_plot, y=value, color=lake_name))+
  labs(x = NULL, y = NULL)+
  geom_point()+
  geom_line(size = 1.2) + 
  facet_wrap(~ name, scales = "free_y", nrow = 2, ncol = 4) +   # 2 rows, 4 columns
  theme(
    legend.position = "none", # remove legend (saved in figures folder)
    panel.background = element_rect(fill = "gray99", color = NA),  # gray panels
    panel.grid.major = element_blank(),
    panel.grid.minor = element_blank(),  # remove grid lines
    plot.title = element_text(face = "bold", size = 10, hjust = 0, margin = margin(b=5)),  # bold, larger, centered, margin below
    strip.text = element_blank(),  # remove default facet labels
    panel.border = element_rect(color = "black", fill = NA, size = 0.8),  # black border
    axis.text.x = element_text(size = 14),   # x-axis ticks
    axis.text.y = element_text(size = 14),    # y-axis ticks
    panel.spacing.x = unit(0.2, "in"),   # increase space between columns
    panel.spacing.y = unit(0.7, "in"),
  ) + 
  scale_color_manual( 
    values = c( 
    "Upper Fourmile Lake" = "#0072B2",  
    "Lower Fourmile Lake" = "yellowgreen",  
    "Turkey Creek Lake" = "#D55E00"))

# save plots (with set dimensions)
ggsave("figures/pigment_flux_plot.png", pigment_flux_plot, width = 14, height = 5, units = "in", dpi = 300) 

# has no units, will go back and add those manually (can check quickly by deleting "strip.text = ..." line)
```

##Pigment conc. plots vs. time
```{R}
# Create pigment plot
pigment_conc_plot <- 
  all_lakes %>%
  pivot_longer(cols = c(bd:year_CE_top, year_CE_bottom:s), names_to = "name", values_to = "value") %>%
  filter(name %in% c("chl_a_nmol_gTC", "pheo_a_nmol_gTC", "PI", "lutein_nmol_gTC", 
                     "diato_nmol_gTC", "allo_nmol_gTC", "echine_nmol_gTC", "cantha_nmol_gTC"
  )) %>% # filter instead of select bc these are rows after using pivot...
  # filter(!is.na(value) & !is.na(depth_top)) %>% # remove missing points to make line continuous
  mutate(name = factor(name, levels = pigment_order), # set facet order 
         lake_name = factor(lake_name, levels = c("Upper Fourmile Lake", 
                                                  "Turkey Creek Lake", 
                                                  "Lower Fourmile Lake"))) %>% # draw lines in order
  ggplot(aes(x=year_CE_plot, y=value, color=lake_name))+
  labs(x = NULL, y = NULL)+
  geom_point()+
  geom_line(size = 1.2) + 
  facet_wrap(~ name, scales = "free_y", nrow = 2, ncol = 4) +   # 2 rows, 4 columns
  theme(
    legend.position = "none", # remove legend (saved in figures folder)
    panel.background = element_rect(fill = "gray99", color = NA),  # gray panels
    panel.grid.major = element_blank(),
    panel.grid.minor = element_blank(),  # remove grid lines
    plot.title = element_text(face = "bold", size = 10, hjust = 0, margin = margin(b=5)),  # bold, larger, centered, margin below
    strip.text = element_blank(),  # remove default facet labels
    panel.border = element_rect(color = "black", fill = NA, size = 0.8),  # black border
    axis.text.x = element_text(size = 14),   # x-axis ticks
    axis.text.y = element_text(size = 14),    # y-axis ticks
    panel.spacing.x = unit(0.2, "in"),   # increase space between columns
    panel.spacing.y = unit(0.7, "in"),
  ) + 
  scale_color_manual( 
    values = c( 
    "Upper Fourmile Lake" = "#0072B2",  # 
    "Lower Fourmile Lake" = "yellowgreen",  # 
    "Turkey Creek Lake" = "#D55E00"))

# save plots (with set dimensions)
ggsave("figures/pigment_conc_plot.png", pigment_conc_plot, width = 14, height = 5, units = "in", dpi = 300) 

# has no units, will go back and add those manually (can check quickly by deleting "strip.text = ..." line)
```
