==> analysis/02d_LMs.Rmd <==
---
title: "02d_LMs"
author: "Julia Pop"
date: "2025-10-11"
output: html_document
---

For UFM you could just run a linear model for those dates. Then, you can just plot the raw data and draw a box with geom_rect() or color the points with dates so that we can visually see the more reliable dated period or use geom_smooth(method="lm", se=TRUE) to plot the linear model fit only on the dated segments (like you did above for the gam). You'll just need to qualitatively describe the patterns in your results and explain in the methods that we don't have reliable dates going back farther, and what you did to extrapolate the dates from a constant sedimentation rate. For example, for d15N you could say something like "For the dated period, d15N declined linearly from a high of 2.8 in 19XX to a low of 2.0 in 20XX (slope & 95% confidence interval, F-statistic). Prior to the dated sediments, d15N deviated between XX and XX for likely several hundred years before declining in modern sediments". If there is no linear "trend" don't bother plotting the predictions or confidence intervals, but do report the null results in the results section. 


```{r setup, include=FALSE}
# code to troublesheet file path issues. 
# By default, knitr uses the folder containing the .Rmd file as
# the working directory. This code allows us to set the folder
# with our .rProj in it as the root. 

# find_root looks upwards from the .Rmd until it finds a file or
# folder that matches the criteria "folder with a .Rproj file in
# it."

knitr::opts_knit$set(root.dir = rprojroot::find_root(rprojroot::is_rstudio_project))
```

#Libraries -- add these here
```{r, echo = FALSE}
source("scripts/libraries.R")
source("scripts/GAM_functions.R")
```

## Pull in data
```{r}
UFM_short_extrap <- read.csv("output/UFM_short_extrap.csv") %>%
  select(-X)
```

# Linear Models
##d15N - signif
```{r}

# Fit a linear model to interpolated dates.
UFM_d15N_lm <- lm(d15N ~ year_CE_plot, data = UFM_short_extrap)
summary(UFM_d15N_lm)

# calculate CIs auto
confint(UFM_d15N_lm, level = 0.95)

# calculate CIs manually
df <- UFM_d15N_lm$df.residual
t_value <- qt(0.975, df)  # 95% CI

summary(UFM_d15N_lm)$coefficients[ , "Std. Error"] # se = 0.0035
lower <- -0.01320 + t_value*0.003490199 # 
upper <- -0.01320 - t_value*0.003490199 # 

cbind(lower, upper)


acf(residuals(UFM_d15N_lm))
# wide CI for zero autocorrelation because n is small
# the test is not very sensitive because the sample size is so small. 
# but still it looks decent. (no consistent trends)

# plot: 
ggplot(UFM_short_extrap) +
  geom_point(aes(year_plot_extrap, d15N)) + # plot all the points
  geom_smooth(                              # plot the model 
    data = UFM_short_extrap %>% filter(!is.na(year_CE_plot)),
    aes(x = year_CE_plot, y = d15N),
    method = "lm", se = TRUE, linewidth = 0.8, col = "black", linetype = "twodash") + 
  ylab(expression(paste(delta^{15}, "N (", "\u2030", ")"))) +
  xlab(expression(paste("Year (", italic("CE"), ")"))) +
  ggtitle("Upper Fourmile Lake") + 
  # adjust gridlines and axis labels
  scale_x_continuous(
  breaks = seq(1400, 2025, by = 50),      # major tick labels every 50 years
  minor_breaks = seq(1400, 2030, by = 10) # grid lines every 10 years
) +
theme_bw() 

ggsave("figures/UFM_d15N.png",
       dpi=600,
       height=5,
       width=8,
       units="in")

```

##d15N - as GAM
```{r}
names(UFM_short_extrap)

### K (sets the maximum # of bases for our spline): 
# Our dataset is still small: rule of thumb is to not let k approach n
    # how much data are we working with? 
    sum(!is.na(UFM_short_extrap$year_CE_plot)) # 10 observations... not so many
    sum(!is.na(UFM_short_extrap$year_plot_extrap)) # 41 observations. 

# DROP dates < 1860 (oldest LFM)
mod_d15N_UFM <- gamm(d15N ~ s(year_plot_extrap, k = 10), 
                     family = gaussian(link = "identity"), 
                     data = UFM_short_extrap %>% filter(year_plot_extrap>1860),
                     method = "REML")

summary(mod_d15N_UFM$lme)

resid_values_d15N_UFM <- residuals(mod_d15N_UFM$lme, type = "normalized")
acf(resid_values_d15N_UFM) # looks good!

## summary object for use in document
d15NSumm_UFM <- summary(mod_d15N_UFM$gam)
d15NSumm_UFM #Gives you the P values, degrees of freedom...

### Check diagnostics. You might need to adjust model

par(mfrow = c(2, 2))   # Set up a 2x2 grid
gam.check(mod_d15N_UFM$gam)
par(mfrow = c(1, 1)) 




### NOW PLOT! ###
N <- 300    # number of points at which to evaluate the splines

# Predict (get 200 points at which to evaluate)
d15NYear_UFM <- with(UFM_short_extrap %>% filter(year_plot_extrap>1860), 
                     data.frame(year_plot_extrap = seq(min(year_plot_extrap, na.rm=TRUE), max(year_plot_extrap, na.rm=TRUE), length.out = 200)))

d15NYear_UFM <- cbind(d15NYear_UFM,
                      data.frame(predict(mod_d15N_UFM$gam, d15NYear_UFM,
                                         type="response", se.fit = TRUE)))

### this calculates on the link scale 
d15NYear_UFM <- transform(d15NYear_UFM, upper = fit + (2 * se.fit), lower = fit - (2 * se.fit))
d15NYear_UFM$lake_ID <- 'UFM'

## Plot fitted trends
ggplot(d15NYear_UFM, aes(x = year_plot_extrap, y = fit)) +
  geom_ribbon(aes(ymin = (lower),
                  ymax = (upper), x = year_plot_extrap),
              alpha = 0.2, inherit.aes = FALSE, fill = "black") +
  geom_point(data = UFM_short_extrap %>% filter(year_plot_extrap>1860), mapping = aes(x = year_plot_extrap, y = d15N), inherit.aes = FALSE) +
  geom_line() 

# year_plot_extrap = year_CE_plot where dates are interpolated. Can use either for plotting points.

#First derivatives using gratia package instead of the 
#functions "helper_functions.R" (that Cale sent)
############
############
fd_inc = confint(fderiv(mod_d15N_UFM))
fd_inc

#Add years for plotting
years <- with(UFM_short_extrap,
                     data.frame(year_plot_extrap = seq(min(year_plot_extrap, na.rm=TRUE), max(year_plot_extrap, na.rm=TRUE), length.out = 200)))

fd_inc <- cbind(fd_inc,years)

#Plot the first derivatives
fd_inc %>%
  select(-term) %>%
  pivot_longer(lower:upper) %>%
  ggplot(aes(x=year_plot_extrap,y=value,linetype=name))+
  geom_line()+
    scale_linetype_manual(values = c(
    "lower" = "dashed", #lower 95% confidence interval
    "upper" = "dashed", #upper 95% confidence interval
    "est"   = "solid"
  ))+
  theme(legend.position="none")+
  geom_hline(yintercept=0, color="red")
# Anywhere the confidence intervals DONT overlap zero is a statistically
# significant acceleration or deceleration in the trend


# When do they occur? You can check with an if/else type statement
fd_inc_timing <- fd_inc %>%
  mutate(change_type = case_when(lower < 0 & upper < 0 ~ "sig. dec.",
                                 lower > 0 & upper > 0 ~ "sig. inc.",
                                 TRUE ~ NA))
# View(fd_inc_timing)
#What you can see if a brief period of decline ~1900-1934, and again from 
#1960-present. This model, relative to Cale's seems more sensitive to that
#slight increase pre-1900. I think Cale's weighting is probably better

# For graphing, we want to join the predicted/fitted data with the 
# information with significant periods of change. Let's call it 'pred'

pred <- left_join(d15NYear_UFM,
                  fd_inc_timing %>% 
                    select(year_plot_extrap, change_type), #select only 2 columns since
                  #lower and upper have two different meanings in these 2 dataframes
                  by="year_plot_extrap") 
  # mutate(fit = ifelse(change_type == "sig. dec.", fit, NA))

# Add a segment ID for plotting, so there there are multiple
# periods of change, ggplot doesn't connect the lines together
pred2 <- pred %>%
  mutate(seg_id = with(rle(ifelse(is.na(change_type), "none", change_type)),
                       rep(seq_along(values), lengths)))

# Keep only rows that are sig. inc or sig. dec
pred2 <- pred2 %>% filter(!is.na(change_type))

# Now make a pretty graph that shows the raw data, fitted trend,
# trend as well as periods of statistically significant increase/decrease!

highlight_data <- tibble(xmin = 1949, xmax = 2025, ymin = -Inf, ymax = Inf)


ggplot(UFM_short_extrap %>% filter(year_plot_extrap>1860)) +
    #Add a box to show the reader the time period for interpretation
  geom_rect(
    mapping = aes(ymin = ymin, ymax = ymax, xmin = xmin, xmax = xmax), 
    data = highlight_data, 
    alpha = 0.1,
    fill = "blue",
    inherit.aes = FALSE
  )+
  geom_point(aes(year_plot_extrap, d15N)) +
  geom_ribbon(data = pred,
              aes(x = year_plot_extrap, ymin = lower, ymax = upper),
              alpha = 0.2) +
  geom_line(data = pred,
            aes(x = year_plot_extrap, y = fit),
            linewidth = 1,
            color = "black") +
  geom_line(data = pred2,
            aes(x = year_plot_extrap, y = fit,
                color = change_type,
                group = seg_id),
            linewidth = 1) +
  scale_color_manual(values = c("sig. inc." = "#CC9933",
                                "sig. dec." = "#2f6cad"),
                     name ="Statistic") +
  # scale_x_continuous(breaks=seq(1880,2025,by=25))+
  ylab(expression(paste(delta^{15}, "N (", "\u2030", ")"))) +
  xlab(expression(paste("Year (", italic("CE"), ")"))) +
  ggtitle("Turkey Creek Lake")

ggsave("figures/UFM_d15N_alt.png",
       dpi=600,
       height=5,
       width=8,
       units="in")
```

##d13C - not signif
```{r}

# Fit a linear model to interpolated dates.
UFM_d13C_lm <- lm(d13C ~ year_CE_plot, data = UFM_short_extrap)
summary(UFM_d13C_lm)

confint(UFM_d13C_lm, level = 0.95)

acf(residuals(UFM_d13C_lm))

# plot: 
ggplot(UFM_short_extrap) +
  geom_point(aes(year_plot_extrap, d13C)) + # plot all the points
  geom_smooth(                              # plot the model 
    data = UFM_short_extrap %>% filter(!is.na(year_CE_plot)),
    aes(x = year_CE_plot, y = d13C),
    method = "lm", se = TRUE, linewidth = 0.8, col = "black", linetype = "twodash") + 
  ylab(expression(paste(delta^{13}, "C (", "\u2030", ")"))) +
  xlab(expression(paste("Year (", italic("CE"), ")"))) +
  ggtitle("Upper Fourmile Lake") + 
  # adjust gridlines and axis labels
  scale_x_continuous(
  breaks = seq(1400, 2025, by = 50),      # major tick labels every 50 years
  minor_breaks = seq(1400, 2030, by = 10) # grid lines every 10 years
) +
theme_bw() 

ggsave("figures/UFM_d13C.png",
       dpi=600,
       height=5,
       width=8,
       units="in")

```

##percent C - signif
```{r}
# Fit a linear model to interpolated dates.
UFM_percent.C_lm <- lm(percent.C ~ year_CE_plot, data = UFM_short_extrap)
summary(UFM_percent.C_lm)

confint(UFM_percent.C_lm, level = 0.95)

acf(residuals(UFM_percent.C_lm))

# plot: 
ggplot(UFM_short_extrap) +
  geom_point(aes(year_plot_extrap, percent.C)) + # plot all the points
  geom_smooth(                              # plot the model 
    data = UFM_short_extrap %>% filter(!is.na(year_CE_plot)),
    aes(x = year_CE_plot, y = percent.C),
    method = "lm", se = TRUE, linewidth = 0.8, col = "black", linetype = "twodash") + 
  ylab("Total Carbon (%)") +
  xlab(expression(paste("Year (", italic("CE"), ")"))) +
  ggtitle("Upper Fourmile Lake") + 
  # adjust gridlines and axis labels
  scale_x_continuous(
  breaks = seq(1400, 2025, by = 50),      # major tick labels every 50 years
  minor_breaks = seq(1400, 2030, by = 10) # grid lines every 10 years
) +
theme_bw() 

ggsave("figures/UFM_percent.C.png",
       dpi=600,
       height=5,
       width=8,
       units="in")
```


##percent N - marginally signif
```{r}

# Fit a linear model to interpolated dates.
UFM_percent.N_lm <- lm(percent.N ~ year_CE_plot, data = UFM_short_extrap)
summary(UFM_percent.N_lm)

confint(UFM_percent.N_lm, level = 0.95)


acf(residuals(UFM_percent.N_lm))

# plot: 
ggplot(UFM_short_extrap) +
  geom_point(aes(year_plot_extrap, percent.N)) + # plot all the points
  geom_smooth(                              # plot the model 
    data = UFM_short_extrap %>% filter(!is.na(year_CE_plot)),
    aes(x = year_CE_plot, y = percent.N),
    method = "lm", se = TRUE, linewidth = 0.8, col = "black", linetype = "twodash") + 
  ylab("Total Nitrogen (%)") +
  xlab(expression(paste("Year (", italic("CE"), ")"))) +
  ggtitle("Upper Fourmile Lake") + 
  # adjust gridlines and axis labels
  scale_x_continuous(
  breaks = seq(1400, 2025, by = 50),      # major tick labels every 50 years
  minor_breaks = seq(1400, 2030, by = 10) # grid lines every 10 years
) +
theme_bw() + 
coord_cartesian(ylim = c(0, 1.1))


ggsave("figures/UFM_percent.N.png",
       dpi=600,
       height=5,
       width=8,
       units="in")
```

##C:N - not signif
```{r}
# Fit a linear model to interpolated dates.
UFM_C.N_lm <- lm(C.N ~ year_CE_plot, data = UFM_short_extrap)
summary(UFM_C.N_lm)

confint(UFM_C.N_lm, level = 0.95)

acf(residuals(UFM_C.N_lm))

# plot: 
ggplot(UFM_short_extrap) +
  geom_point(aes(year_plot_extrap, C.N)) + # plot all the points
  geom_smooth(                              # plot the model 
    data = UFM_short_extrap %>% filter(!is.na(year_CE_plot)),
    aes(x = year_CE_plot, y = C.N),
    method = "lm", se = TRUE, linewidth = 0.8, col = "black", linetype = "twodash") + 
  ylab("C:N ratio") +
  xlab(expression(paste("Year (", italic("CE"), ")"))) +
  ggtitle("Upper Fourmile Lake") + 
  # adjust gridlines and axis labels
  scale_x_continuous(
  breaks = seq(1400, 2025, by = 50),      # major tick labels every 50 years
  minor_breaks = seq(1400, 2030, by = 10) # grid lines every 10 years
) +
theme_bw() 

ggsave("figures/UFM_C.N.png",
       dpi=600,
       height=5,
       width=8,
       units="in")
```

##Pheo a - not signif
```{r}

# Fit a linear model to interpolated dates.
UFM_pheoa_lm <- lm(pheo_a_nmol_gTC ~ year_CE_plot, data = UFM_short_extrap)
summary(UFM_pheoa_lm)

confint(UFM_pheoa_lm, level = 0.95)


acf(residuals(UFM_pheoa_lm))

# plot: 
ggplot(UFM_short_extrap) +
  geom_point(aes(year_plot_extrap, pheo_a_nmol_gTC)) + # plot all the points
  geom_smooth(                              # plot the model 
    data = UFM_short_extrap %>% filter(!is.na(year_CE_plot)),
    aes(x = year_CE_plot, y = pheo_a_nmol_gTC),
    method = "lm", se = TRUE, linewidth = 0.8, col = "black", linetype = "twodash") + 
  ylab(expression("pheophytin a ("*nmol~g^-1~"total carbon"*")")) + 
  xlab(expression(paste("Year (", italic("CE"), ")"))) +
  ggtitle("Upper Fourmile Lake") + 
  # adjust gridlines and axis labels
  scale_x_continuous(
  breaks = seq(1400, 2025, by = 50),      # major tick labels every 50 years
  minor_breaks = seq(1400, 2030, by = 10)) +  # grid lines every 10 years
theme_bw() + 
  coord_cartesian(ylim = c(0, 1500))


ggsave("figures/UFM_pheoa.png",
       dpi=600,
       height=5,
       width=8,
       units="in")
```

##Greens - not at all signif
```{r}

# Fit a linear model to interpolated dates.
UFM_greens_lm <- lm(greens ~ year_CE_plot, data = UFM_short_extrap)
summary(UFM_greens_lm)

confint(UFM_greens_lm, level = 0.95)

acf(residuals(UFM_greens_lm))

# plot: 
ggplot(UFM_short_extrap) +
  geom_point(aes(year_plot_extrap, greens)) + # plot all the points
  geom_smooth(                              # plot the model 
    data = UFM_short_extrap %>% filter(!is.na(year_CE_plot)),
    aes(x = year_CE_plot, y = greens),
    method = "lm", se = TRUE, linewidth = 0.8, col = "black", linetype = "twodash") + 
  ylab(expression("pheo b & lutein ("*nmol~g^-1~"total carbon"*")")) + 
  xlab(expression(paste("Year (", italic("CE"), ")"))) +
  ggtitle("Upper Fourmile Lake") + 
  # adjust gridlines and axis labels
  scale_x_continuous(
  breaks = seq(1400, 2025, by = 50),      # major tick labels every 50 years
  minor_breaks = seq(1400, 2030, by = 10)) +  # grid lines every 10 years
theme_bw() + 
  coord_cartesian(ylim = c(0, 1500))


ggsave("figures/UFM_greens.png",
       dpi=600,
       height=5,
       width=8,
       units="in")
```

##Diatoxanthin - not at all signif
```{r}

# Fit a linear model to interpolated dates.
UFM_diato_lm <- lm(diato_nmol_gTC ~ year_CE_plot, data = UFM_short_extrap)
summary(UFM_diato_lm)

confint(UFM_diato_lm, level = 0.95)


acf(residuals(UFM_diato_lm))

# plot: 
ggplot(UFM_short_extrap) +
  geom_point(aes(year_plot_extrap, diato_nmol_gTC)) + # plot all the points
  geom_smooth(                              # plot the model 
    data = UFM_short_extrap %>% filter(!is.na(year_CE_plot)),
    aes(x = year_CE_plot, y = diato_nmol_gTC),
    method = "lm", se = TRUE, linewidth = 0.8, col = "black", linetype = "twodash") + 
  ylab(expression("diatoxanthin ("*nmol~g^-1~"total carbon"*")")) + 
  xlab(expression(paste("Year (", italic("CE"), ")"))) +
  ggtitle("Upper Fourmile Lake") + 
  # adjust gridlines and axis labels
  scale_x_continuous(
  breaks = seq(1400, 2025, by = 50),      # major tick labels every 50 years
  minor_breaks = seq(1400, 2030, by = 10)) +  # grid lines every 10 years
theme_bw() + 
  coord_cartesian(ylim = c(0, 150))


ggsave("figures/UFM_diato.png",
       dpi=600,
       height=5,
       width=8,
       units="in")
```

##Alloxanthin - not signif
```{r}

# Fit a linear model to interpolated dates.
UFM_allo_lm <- lm(allo_nmol_gTC ~ year_CE_plot, data = UFM_short_extrap)
summary(UFM_allo_lm)

confint(UFM_allo_lm, level = 0.95)


acf(residuals(UFM_allo_lm))

# plot: 
ggplot(UFM_short_extrap) +
  geom_point(aes(year_plot_extrap, allo_nmol_gTC)) + # plot all the points
  geom_smooth(                              # plot the model 
    data = UFM_short_extrap %>% filter(!is.na(year_CE_plot)),
    aes(x = year_CE_plot, y = allo_nmol_gTC),
    method = "lm", se = TRUE, linewidth = 0.8, col = "black", linetype = "twodash") + 
  ylab(expression("alloxanthin ("*nmol~g^-1~"total carbon"*")")) + 
  xlab(expression(paste("Year (", italic("CE"), ")"))) +
  ggtitle("Upper Fourmile Lake") + 
  # adjust gridlines and axis labels
  scale_x_continuous(
  breaks = seq(1400, 2025, by = 50),      # major tick labels every 50 years
  minor_breaks = seq(1400, 2030, by = 10)) +  # grid lines every 10 years
theme_bw() + 
  coord_cartesian(ylim = c(0, 225))

ggsave("figures/UFM_allo.png",
       dpi=600,
       height=5,
       width=8,
       units="in")
```

##Cyanos - not signif
```{r}

# Fit a linear model to interpolated dates.
UFM_cyanos_lm <- lm(cyanos ~ year_CE_plot, data = UFM_short_extrap)
summary(UFM_cyanos_lm)

confint(UFM_cyanos_lm, level = 0.95)


acf(residuals(UFM_cyanos_lm))

# plot: 
ggplot(UFM_short_extrap) +
  geom_point(aes(year_plot_extrap, cyanos)) + # plot all the points
  geom_smooth(                              # plot the model 
    data = UFM_short_extrap %>% filter(!is.na(year_CE_plot)),
    aes(x = year_CE_plot, y = cyanos),
    method = "lm", se = TRUE, linewidth = 0.8, col = "black", linetype = "twodash") + 
  ylab(expression("echinenone & canthaxanthin ("*nmol~g^-1~"total carbon"*")")) + 
  xlab(expression(paste("Year (", italic("CE"), ")"))) +
  ggtitle("Upper Fourmile Lake") + 
  # adjust gridlines and axis labels
  scale_x_continuous(
  breaks = seq(1400, 2025, by = 50),      # major tick labels every 50 years
  minor_breaks = seq(1400, 2030, by = 10)) +  # grid lines every 10 years
theme_bw() + 
  coord_cartesian(ylim = c(0, 150))


ggsave("figures/UFM_cyanos.png",
       dpi=600,
       height=5,
       width=8,
       units="in")
```
